# Swagger/swagger-ui/swagger.yml
swagger: '2.0'
info:
  version: "0.0.1"
  title: Cruise-Service
# during dev, should point to your local machine
host: localhost:8000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: api_Key
    in: header
    # scheme: bearer
    # bearerFormat: JWT
paths:
...
  /findUserById:
       # binds a127 app logic to a route
    x-swagger-router-controller: crud
    get:
      description: Returns User Info by Id
      # used as the method name of the controller
      operationId: findUserById
      parameters:
        - name: uid
          in: query
          description: Enter User ID
          required: true
          type: integer
      security:
        - BearerAuth: []
      x-security-scopes:
        - admin
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/getUserInfoById"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  getUserInfoById:
    required:
      - uid
    properties:
      uid:
         type: integer
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string